refactor: unify ingestion module
